ai-support-agent/requirements.txt
streamlit
faiss-cpu
sentence-transformers
python-dotenv
numpy
google-generativeai
